+++ src/apis/postApi.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { serverURL } from './endpoints';
+import { getCookie } from '../utils/cookie';
 
 export const getPostByChannelApi = async channelId => {
   try {
@@ -31,8 +32,7 @@ export const getChannelNameByIdApi = async channelId => {
 };
 
 export const deletePostApi = async contentID => {
-  const token =
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7Il9pZCI6IjY3OTA4ZGFlZThhMWU0MzQ5ZWQ3NmVjMiIsImVtYWlsIjoidGVzdCJ9LCJpYXQiOjE3Mzc2OTk2ODF9.RCuyIwJ-9wIuR_gYYQc4gaLwjsmletQQMiWqKL9h5ug';
+  const token = getCookie('jwt');
   try {
     const response = await axios.delete(`${serverURL}/posts/delete`, {
       headers: {
@@ -41,7 +41,6 @@ export const deletePostApi = async contentID => {
       },
       data: { id: contentID },
     });
-
     return response.data;
   } catch (error) {
     console.error('게시글 삭제에 실패했습니다.', error);
