+++ src/pages/MyPage/components/main/post/PostsSection.jsx
@@ -2,9 +2,10 @@
 import { useQuery } from '@tanstack/react-query';
 import Post from './Post';
 import { getUserPost } from '../../../../../apis/getUserData';
+import { useSelector } from 'react-redux';
 
 const PostsSection = () => {
-  const userId = '67908daee8a1e4349ed76ec2'; // 더미 데이터
+  const { userId } = useSelector(state => state.user);
 
   const { data, isLoading, isError, error } = useQuery({
     queryKey: [userId],
